Implement the message handler

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/networking/libp2p/rpc/BlobsSidecarsByRangeRequestMessage.java
@@ -66,6 +66,6 @@ public class BlobsSidecarsByRangeRequestMessage
 
   @Override
   public int getMaximumRequestChunks() {
-    return Math.toIntExact(getCount().longValue());
+    return getCount().intValue();
   }
 }
